Initial Migration to custom resources
