+++ metadata.rb
@@ -7,11 +7,14 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 source_url       'https://github.com/sous-chefs/chef-grafana'
 issues_url       'https://github.com/sous-chefs/chef-grafana/issues'
 chef_version     '>= 13.0'
-version          '3.0.1'
+version          '4.0.0'
 
-supports 'ubuntu', '>= 16.04'
-supports 'debian', '>= 8.0'
-supports 'centos', '>= 6.4'
-supports 'redhat', '>= 6.4'
+supports 'debian'
+supports 'ubuntu'
+supports 'redhat'
+supports 'centos'
+supports 'fedora'
+supports 'amazon'
+supports 'scientific'
 
 depends 'nginx', '>= 7.0'
