+++ resources/plugin.rb
@@ -1,7 +1,41 @@
-actions :install, :update, :remove
+# Resource properties
+property :grafana_cli_bin, kind_of: String, required: false, default: '/usr/sbin/grafana-cli'
+
 default_action :install
 
-# Resource properties
-attribute :grafana_cli_bin, kind_of: String, required: false, default: '/usr/sbin/grafana-cli'
+action :install do
+  plugin_name = new_resource.name
+  binary = new_resource.grafana_cli_bin
+  raise "#{plugin_name} is not available" unless ::GrafanaCookbook::Plugin.available?(plugin_name, binary)
+  execute "Installing plugin #{plugin_name}" do
+    command ::GrafanaCookbook::Plugin.build_cli_cmd(plugin_name, 'install', binary)
+    not_if { current_resource.installed }
+  end
+end
 
-attr_accessor :installed
+action :update do
+  plugin_name = new_resource.name
+  binary = new_resource.grafana_cli_bin
+  if current_resource.installed
+    execute "Updating plugin #{plugin_name}" do
+      command ::GrafanaCookbook::Plugin.build_cli_cmd(plugin_name, 'update', binary)
+    end
+  else
+    Chef::Log.warn "Impossible to upgrade plugin #{plugin_name} because it is not installed. We will install it."
+    run_action(:install)
+  end
+end
+
+action :remove do
+  plugin_name = new_resource.name
+  binary = new_resource.grafana_cli_bin
+  execute "Removing plugin #{name}" do
+    command ::GrafanaCookbook::Plugin.build_cli_cmd(plugin_name, 'remove', binary)
+    only_if { current_resource.installed }
+  end
+end
+
+def load_current_resource
+  @current_resource = Chef::Resource.resource_for_node(:grafana_plugin, node).new(@new_resource.name)
+  @current_resource.installed = GrafanaCookbook::Plugin.installed?(new_resource.name, new_resource.grafana_cli_bin)
+end
