+++ spec/unit/nginx_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe 'grafana::_nginx' do
   cached(:chef_run) do
-    ChefSpec::SoloRunner.new(platform: 'centos').converge described_recipe
+    ChefSpec::SoloRunner.new(platform: 'centos', :version => '7.5.1804').converge described_recipe
   end
 
   it 'enables a nginx site' do
